shuffle some code around, rename to --region
