+++ remarkable_mouse/pynput.py
@@ -1,7 +1,8 @@
 import logging
 import struct
-import sys
-from screeninfo import get_monitors, Monitor
+from screeninfo import get_monitors
+
+from .common import get_monitor
 
 logging.basicConfig(format='%(message)s')
 log = logging.getLogger(__name__)
@@ -26,62 +27,6 @@ wacom_height = 20967
 # touchscreen dimensions
 # finger_width = 767
 # finger_height = 1023
-
-
-# Pop up a window, ask the user to move the window, and then get the position of the window's contents
-def ask_user_for_position(args):
-    try:
-        import tkinter as tk
-        from tkinter import ttk
-    except ImportError:
-        print(
-            "Unable to import tkinter; please follow the instructions at https://tkdocs.com/tutorial/install.html to install it")
-        sys.exit(1)
-
-    window = tk.Tk()
-
-    # A bit of an ugly hack to get this function to run synchronously
-    # Ideally would use full async support, but this solution required minimal changes to rest of code
-    selected_pos = None
-
-    def on_click():
-        nonlocal selected_pos
-        selected_pos = Monitor(window.winfo_x(), window.winfo_y(), window.winfo_width(), window.winfo_height(),
-                               name="Fake monitor from rect selection")
-        window.destroy()
-
-    confirm = ttk.Button(window, text="Drag and resize this button to the desired mouse range, then click",
-                         command=on_click)
-    confirm.grid(column=0, row=0, sticky=(tk.N, tk.S, tk.E, tk.W))
-
-    window.columnconfigure(0, weight=1)
-    window.rowconfigure(0, weight=1)
-
-    window.attributes('-alpha', 0.5)
-    window.title("Remarkable Mouse")
-
-    if args.orientation == 'bottom' or args.orientation == 'top':
-        window.geometry("702x936")
-    else:
-        window.geometry("936x702")
-
-    window.mainloop()
-
-    if selected_pos is None:
-        log.debug("Window closed without giving mouse range")
-        sys.exit(1)
-
-    return selected_pos
-
-
-# get info of where we want to map the tablet to
-def get_screen_info(args):
-    if args.rect:
-        monitor = ask_user_for_position(args)
-    else:
-        monitor = get_monitors()[args.monitor]
-
-    return monitor
 
 
 # remap wacom coordinates to screen coordinates
@@ -129,7 +74,7 @@ def read_tablet(args, remote_device):
 
     mouse = Controller()
 
-    monitor = get_screen_info(args)
+    monitor = get_monitor(args)
     log.debug('Chose monitor: {}'.format(monitor))
 
     while True:
